[rocprofiler-compute] Fold kernel name helpers into the shared test module
- Move normalize_kernel_name/normalize_kernel_names into tests/common.py and
  drop the single-consumer module along with its test target
- Bugfix: guard the torch-trace stats report in launch.py so a failure there
  cannot replace the workload's own exception
- State the supported PyTorch versions once in the profile guide, and keep the
  changelog entry to what a user acts on
