+++ projects/rocprofiler-compute/CHANGELOG.md
@@ -16,11 +16,7 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
   * Each edge now reports the traffic measured at the interface it represents.
   * Updated arrows, labels, and the legend in the memory chart to better represent their meaning.
 
-* `--torch-trace` requires PyTorch 2.13 or 2.14.
-
-* `--torch-trace` loads a CMake-built `torch_trace_collector-<torch-version>.so`
-  instead of compiling a helper under `~/.cache/rocprofiler-compute/`. A
-  workload without a matching module fails.
+* `--torch-trace` now requires PyTorch 2.13 or 2.14, installed alongside ROCm.
 
 * Redesigned the CDNA (gfx9) Memory Chart with a new Rich-based layout that improves readability in the terminal. Added Non-buffer/Buffer request breakdowns (Read/Write/Atomic wavefronts) and L2-Fabric bandwidth metrics across all CDNA architectures.
   * gfx908–gfx942: added HBM and remote traffic percentages.
