+++ projects/rocprofiler-compute/CMakeLists.txt
@@ -876,14 +876,6 @@ add_test(
 )
 
 add_test(
-    NAME test_torch_trace_kernel_names
-    COMMAND
-        ${PYTHON_TEST_COMMAND} -m pytest
-        --junitxml=tests/test_torch_trace_kernel_names.xml ${COV_OPTION}
-        tests/unit/test_torch_trace_kernel_names.py ${WORKING_DIR_OPTION}
-)
-
-add_test(
     NAME test_inject_roctx_package
     COMMAND
         ${PYTHON_TEST_COMMAND} -m pytest --junitxml=tests/test_inject_roctx_package.xml
@@ -1072,7 +1064,6 @@ if(${ENABLE_COVERAGE})
         test_profile_torch_trace
         test_profile_triton_trace
         test_torch_cpp_loader
-        test_torch_trace_kernel_names
         test_inject_roctx_package
         test_analysis_db
         test_analysis_base
