+++ projects/rocprofiler-compute/tests/test_categories.yaml
@@ -54,7 +54,6 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_kernel_names
       - test_conftest
       - test_mem_chart_gfx9
       - test_mem_chart_gfx11
@@ -142,7 +141,6 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
@@ -236,7 +234,6 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
@@ -330,7 +327,6 @@ test_categories:
       - test_csv_compression
       - test_inject_roctx_package
       - test_torch_cpp_loader
-      - test_torch_trace_kernel_names
       - test_profiler_base
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
